% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanbanR.R
\name{kanbanR}
\alias{kanbanR}
\title{Kanban Board Widget}
\usage{
kanbanR(
  data,
  deleteButtonStyle = list(color = "white", backgroundColor = "red", icon = "🗑️"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A list representing the Kanban board structure.}

\item{width}{The width of the widget. Must be a valid CSS unit or \code{NULL}.}

\item{height}{The height of the widget. Must be a valid CSS unit or \code{NULL}.}

\item{elementId}{An optional ID for the widget element.}
}
\description{
Creates a Kanban Board for visualization and interaction.
}
