% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanbanR.R
\name{kanbanR}
\alias{kanbanR}
\title{Create a Kanban Board}
\usage{
kanbanR(
  data,
  styleOptions = list(headerBg = "#007bff", headerColor = "#fff", headerFontSize =
    "1rem", listNameFontSize = "1rem", cardTitleFontSize = "1rem", deleteList =
    list(backgroundColor = "#007bff", color = "#fff", icon = bs_icon("x")), deleteCard =
    list(backgroundColor = "#007bff", color = "#fff", icon = bs_icon("trash")),
    addButtonText = "Add", cancelButtonText = "Cancel", addCardButtonText = "Add Card",
    cancelCardButtonText = "Cancel"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Named list for the board}

\item{styleOptions}{Named list of style configs:
\itemize{
\item headerBg, headerColor, headerFontSize
\item listNameFontSize
\item cardTitleFontSize
\item deleteList = list(backgroundColor, color, icon)
\item deleteCard = list(backgroundColor, color, icon)
\item addButtonText, cancelButtonText, addCardButtonText, cancelCardButtonText
}}

\item{width, height}{widget dimension}

\item{elementId}{DOM id}
}
\value{
An htmlwidget
}
\description{
Create a Kanban Board
}
